# doc/api.yaml
openapi: 3.0.0
info:
  description: WASAText app backend REST API
  title: WASAText API
  version: 1.0.0
servers:
  - url: http://localhost:3000/v1
...
security:
  - bearerAuth: []
paths:
...
  /users/{username}/conversations/{conversation_id}/messages/{message_id}/comments:
    parameters:
      - $ref: "#/components/parameters/username"
      - $ref: "#/components/parameters/conversation_id"
      - $ref: "#/components/parameters/message_id"
    post:
      tags: ["Comment"]
      summary: "Add an emoji reaction to a message."
      description: "Leave an emoji reaction on another user's message."
      operationId: commentMessage
      requestBody:
        description: "The emoji reaction to be added to the message."
        required: true
        content:
          application/json:
            schema:
              type: object
              description: "Represents the data required to add an emoji reaction to a message."
              properties:
                emoji:
                  type: string
                  description: "The emoji reaction to add to the message."
                  example: "🌈"
                  minLength: 1
                  maxLength: 1
                  ##pattern: "^[\ud83c[\ud000-\udfff]|\ud83d[\ud000-\udfff]|\ud83e[\ud000-\udfff]]+$"
              required:
                - emoji
      responses:
        "200":
          description: "Emoji reaction added successfully."
          content:
            application/json:
              schema:
                type: object
                description: "Represents a response indicating the success of an operation, such as adding an emoji reaction."
                properties:
                  message:
                    type: string
                    description: "A message confirming the successful completion of the operation."
                    example: "Emoji reaction added successfully."
                    minLength: 10
                    maxLength: 100
                    pattern: "^[a-zA-Z0-9 .,!?']+$"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  
  ##uncommentMessage
    delete:
      tags: ["Comment"]
      summary: "Delete an emoji reaction from a message."
      description: "Delete a specific emoji reaction from a message."
      operationId: uncommentMessage
      responses:
        "204":
          description: "Comment deleted successfully."
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }

  ##setGroupPhoto
  /users/{username}/groups/{group_id}/photo:
    parameters:
      - $ref: "#/components/parameters/username"
      - $ref: "#/components/parameters/group_id"
    put:
      tags: ["Group"]
      summary: "Upload group picture."
      description: "Allows an admin to upload a picture for a group they manage."
      operationId: setGroupPhoto
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              description: "Represents a request to upload a new picture for the group."
              properties:
                photo:
                  type: string
                  format: binary
                  description: "The picture to upload for the group."
                  minLength: 10
                  maxLength: 100
      responses:
        "200":
          description: "Group picture uploaded."
          content:
            application/json:
              schema:
                type: object
                description: "Represents a response indicating the success of an operation, such as uploading a group picture."
                properties:
                  message:
                    type: string
                    description: "A message confirming the successful upload of the group picture."
                    example: "Group picture uploaded successfully."
                    minLength: 10
                    maxLength: 100
                    pattern: "^[a-zA-Z0-9 .,!?']+$"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }

  ## setGroupName
  /users/{username}/groups/{group_id}/name:
    parameters:
      - $ref: "#/components/parameters/username"
      - $ref: "#/components/parameters/group_id"
    put:
      tags: ["Group"]
      summary: "Update group name."
      description: "Allows an admin to update the name of a group they manage."
      operationId: setGroupName
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: "Represents a request to update the name of an existing group."
              properties:
                name:
                  type: string
                  description: "The new name for the group."
                  example: "My New Group Name"
                  minLength: 3
                  maxLength: 50
                  pattern: "^[a-zA-Z0-9 ]+$"
      responses:
        "200":
          description: "Group name updated."
          content:
            application/json:
              schema:
                type: object
                description: "Represents a response indicating the success of an operation, such as updating the group name."
                properties:
                  message:
                    type: string
                    description: "A message confirming the successful update of the group name."
                    example: "Group name updated successfully."
                    minLength: 10
                    maxLength: 100
                    pattern: "^[a-zA-Z0-9 .,!?']+$"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }

  ##createGroup
  /users/{username}/groups:
    parameters:
      - $ref: "#/components/parameters/username"
    post:
      tags: ["Group"]
      summary: "Create a new group."
      description: "Create a new group chat with the logged-in user as the admin."
      operationId: createGroup
      requestBody:
        description: "Details of the new group."
        required: true
        content:
          application/json:
            schema:
              type: object
              description: "Represents the data required to create or update a group, including the group’s name, description, and list of members."
              properties:
                group_name:
                  type: string
                  description: "The name of the group."
                  example: "The Clique"
                  minLength: 3
                  maxLength: 50
                  pattern: "^[a-zA-Z0-9 ]+$"

                description:
                  type: string
                  description: "A short description of the group."
                  example: "Group for our band."
                  minLength: 5
                  maxLength: 200
                  pattern: "^[a-zA-Z0-9 .]+$"

                members:
                  type: array
                  items:
                    type: string
                    minLength: 3
                    maxLength: 30
                    pattern: "^[a-zA-Z0-9 ]+$"
                  description: "List of usernames to add to the group."
                  example: ["Anna", "Billie", "Karl"]
                  minItems: 1
                  maxItems: 50
              required:
                - group_name
                - members
      responses:
        "201":
          description: "Group created successfully."
          content:
            application/json:
              schema:
                type: object
                description: "Represents the response after successfully creating a new group, including the group's unique identifier and name."
                properties:
                  group_id:
                    type: string
                    description: "The unique identifier of the newly created group."
                    example: "group12345"
                    minLength: 8
                    maxLength: 50
                    pattern: "^[a-zA-Z0-9]+$"

                  group_name:
                    type: string
                    description: "The name of the created group."
                    example: "The Clique"
                    minLength: 3
                    maxLength: 50
                    pattern: "^[a-zA-Z0-9 ]+$"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409":
          description: "A group with the same name already exists."
          content:
            application/json:
              schema:
                type: object
                description: "Represents a response indicating an error or conflict, such as when a group name is already taken."
                properties:
                  message:
                    type: string
                    description: "A descriptive message stating that the group name is already in use."
                    example: "Group name already in use."
                    minLength: 10
                    maxLength: 100
                    pattern: "^[a-zA-Z0-9 .]+$"
        "500": { $ref: "#/components/responses/InternalServerError" }
    
  ##addToGroup
  /users/{username}/groups/{group_id}/members:
    parameters:
      - $ref: "#/components/parameters/username"
      - $ref: "#/components/parameters/group_id"
    post:
      tags: ["Group"]
      summary: "Add a member to a group."
      description: "Add a specified user to an existing group."
      operationId: addToGroup
      requestBody:
        description: "The username of the member to be added to the group."
        required: true
        content:
          application/json:
            schema:
              type: object
              description: "Represents a request to add a new member to the system or group."
              properties:
                new_member_username:
                  type: string
                  description: "The username of the new member to add."
                  example: "slow_koala"
                  minLength: 3
                  maxLength: 30
                  pattern: "^[a-zA-Z0-9_]+$"
              required:
                - new_member_username
      responses:
        "200":
          description: "Member added successfully."
          content:
            application/json:
              schema:
                type: object
                description: "Represents a response indicating the result of a member-related action."
                properties:
                  message:
                    type: string
                    description: "A confirmation message indicating that the member was added successfully."
                    example: "Member added successfully."
                    minLength: 10
                    maxLength: 100
                    pattern: "^[a-zA-Z0-9 .]+$"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }

  ##leaveGroup
  /users/{username}/groups/{group_id}/members/{member_username}:
    parameters:
      - $ref: "#/components/parameters/username"
      - $ref: "#/components/parameters/group_id"
      - name: member_username
        in: path
        required: true
        schema:
          type: string
          minLength: 3
          maxLength: 30
          pattern: "^[a-zA-Z0-9_]+$"
        description: "The username of the member who wants to leave the group."
    delete:
      tags: ["Group"]
      summary: "Leave a group."
      description: "Allows a user to leave an existing group."
      operationId: leaveGroup
      responses:
        "204":
          description: "User successfully left the group."
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "User is already a member of the group."
          content:
            application/json:
              schema:
                type: object
                description: "Represents an error or informational response for group-related actions."
                properties:
                  message:
                    type: string
                    description: "A detailed message indicating the error or status of the request."
                    example: "User is already in the group."
                    minLength: 10
                    maxLength: 100
                    pattern: "^[a-zA-Z0-9. ]+$"
        "500": { $ref: "#/components/responses/InternalServerError" }
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    BadRequest:
      description: "Invalid/malformed request."
      content:
        application/json:
          schema:
            type: object
            description: "An error response indicating the request was malformed or invalid."
            properties:
              message:
                type: string
                description: "An error response indicating the request was malformed or invalid."
                example: "The request body is malformed."
                minLength: 10
                maxLength: 200
                pattern: "^[A-Z].*\\.$"
    Unauthorized:
      description: "Authentication required."
      content:
        application/json:
          schema:
            type: object
            description: "An error response indicating that the user needs to authenticate."
            properties:
              message:
                type: string
                description: "An error response indicating that the user needs to authenticate."
                example: "Authentication is required to access this resource."
                minLength: 10
                maxLength: 200
                pattern: "^[A-Z].*\\.$"
    InternalServerError:
      description: "Server currently unavailable."
      content:
        application/json:
          schema:
            type: object
            description: "An error response indicating an internal server error."
            properties:
              message:
                type: string
                description: "An error response indicating an internal server error."
                example: "The server is currently unavailable, please try again later."
                minLength: 10
                maxLength: 200
                pattern: "^[A-Z].*\\.$"
    NotFound:
      description: "Resource not found."
      content:
        application/json:
          schema:
            type: object
            description: "An error response indicating the requested resource could not be found."
            properties:
              message:
                type: string
                description: "An error response indicating the requested resource could not be found."
                example: "The requested resource could not be found."
                minLength: 10
                maxLength: 200
                pattern: "^[A-Z].*\\.$"
    Forbidden:
      description: "The operation is forbidden."
      content:
        application/json:
          schema:
            type: object
            description: "An error response indicating that the operation is not allowed."
            properties:
              message:
                type: string
                description: "An error response indicating that the operation is not allowed."
                example: "You do not have permission to perform this operation."
                minLength: 10
                maxLength: 200
                pattern: "^[A-Z].*\\.$"
  parameters:
    username:
      name: username
      in: path
      required: true
      description: "Alias for the User."
      schema:
        type: string
        minLength: 6
        maxLength: 30
        pattern: "^[a-zA-Z0-9_]{6,30}$"
        example: booklover
    conversation_id:
      name: conversation_id
      in: path
      required: true
      description: "ID for a conversation."
      schema:
        type: string
        minLength: 6
        maxLength: 30
        example: convo123
        pattern: "^[a-zA-Z0-9_]+$"
    message_id:
      name: message_id
      in: path
      required: true
      description: "ID for a message."
      schema:
        type: string
        minLength: 6
        maxLength: 30
        example: message123
        pattern: "^[a-zA-Z0-9_]+$"
    group_id:
      name: group_id
      in: path
      required: true
      description: "ID for a group."
      schema:
        type: string
        minLength: 6
        maxLength: 30
        example: group123
        pattern: "^[a-zA-Z0-9_]+$"